issue-1: more reliable out-of-the-box.
The initial issue was that ``configure.py`` needed the settings,
but the underlying problem was deeper. When no social providers
were configured, the login template would fail, forcing users to
set up a social app provider.

The template already had conditional logic to render portions only
when the apps were configured, but that meant changes to ``settings.py``.
So I moved ``settings.py`` to ``settings.template.py`` and update it
in ``make configure``, enabling or disabling Facebook and Google
support appropriately.

Because ``settings.py`` is no longer present when ``make configure`` is
run, it must work without any settings, which means there is no longer
any need to specify settings when invoking it.

Other fixes and changes including:
- ``make clean`` handles spaces in pathnames (e.g. nested python install)
- a ``make veryclean`` target that should get rid of all generated files.
- Changed a lot of the README.
- removed unnecessary modules from settings and ``requirements.txt``.
- when there are no social providers, the login form uses the full width.